Feat: "[레벨3] double 타입의 값을 전달 받아도 연산이 수행하도록 만들기"
- 피연산자를 여러 타입으로 받을 수 있도록 기능을 확장
- 잘못된 연산자를 입력했을 때 처리하도록 수정

Closes: #3

+++ src/calculator/advanced/OperatorType.java
@@ -2,22 +2,22 @@ package calculator.advanced;
 
 public enum OperatorType {
     ADD('+'){
-        public int apply(int firstNumber, int secondNumber){
+        public double apply(double firstNumber, double secondNumber){
             return firstNumber + secondNumber;
         }
     },
     SUBTRACT('-'){
-        public int apply(int firstNumber, int secondNumber){
+        public double apply(double firstNumber, double secondNumber){
             return firstNumber - secondNumber;
         }
     },
     MULTIPLY('*'){
-        public int apply(int firstNumber, int secondNumber){
+        public double apply(double firstNumber, double secondNumber){
             return firstNumber * secondNumber;
         }
     },
     DIVIDE('/'){
-        public int apply(int firstNumber, int secondNumber){
+        public double apply(double firstNumber, double secondNumber){
             if (secondNumber != 0) return firstNumber / secondNumber;
             else throw new ArithmeticException("0으로 나눌 수 없습니다.");
         }
@@ -40,5 +40,5 @@ public enum OperatorType {
         throw new IllegalArgumentException("올바르지 않은 연산자입니다: " + operator);
     }
 
-    public abstract int apply(int firstNumber, int secondNumber);
+    public abstract double apply(double firstNumber, double secondNumber);
 }
